create display folder
